Move tooltip tests to integration tests
They basically spin up the entire server so it's almost end-to-end,
after all.
